+++ data_pipeline/generic_web_api/generic_web_api_data_etl.py
@@ -5,13 +5,14 @@ from pathlib import Path
 import json
 from json.decoder import JSONDecodeError
 
-from typing import Iterable, List
-import dateparser
 from botocore.exceptions import ClientError
 
+from data_pipeline.generic_web_api.transform_data import (
+    ModuleConstant, process_downloaded_data,
+    get_dict_values_from_hierarchy_as_list
+)
 from data_pipeline.utils.data_store.s3_data_service import (
     download_s3_object_as_string,
-    download_s3_json_object,
     upload_s3_object
 )
 from data_pipeline.utils.data_store.bq_data_service import (
@@ -19,51 +20,19 @@ from data_pipeline.utils.data_store.bq_data_service import (
     create_or_extend_table_schema
 
 )
-from data_pipeline.crossref_event_data.etl_crossref_event_data_util import (
-    convert_bq_schema_field_list_to_dict,
-    standardize_field_name,
-    requests_retry_session
-)
-from data_pipeline.utils.pipeline_file_io import iter_write_jsonl_to_file
+from data_pipeline.utils.web_api import requests_retry_session
 
 from data_pipeline.generic_web_api.generic_web_api_config import (
     WebApiConfig
 )
 from data_pipeline.generic_web_api.helper import (
-    ResponseHierarchyKey, UrlComposeParam
+    UrlComposeParam
 )
 from data_pipeline.utils.data_pipeline_timestamp import (
-    get_current_timestamp_as_string
+    get_current_timestamp_as_string,
+    datetime_to_string,
+    parse_timestamp_from_str
 )
-
-
-# pylint: disable=too-few-public-methods
-class ModuleConstant:
-    DEFAULT_TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S%z"
-    DATA_IMPORT_TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S"
-    BQ_SCHEMA_FIELD_NAME_KEY = "name"
-    BQ_SCHEMA_SUBFIELD_KEY = "fields"
-    BQ_SCHEMA_FIELD_TYPE_KEY = "type"
-
-
-def get_timestamp_as_string(
-        timestamp: datetime,
-        timestamp_format: str = ModuleConstant.DEFAULT_TIMESTAMP_FORMAT):
-    return timestamp.strftime(
-        timestamp_format
-    )
-
-
-def parse_timestamp_from_str(timestamp_as_str, time_format: str = None):
-    if time_format:
-        timestamp_obj = datetime.strptime(
-            timestamp_as_str.strip(), time_format
-        )
-    else:
-        timestamp_obj = dateparser.parse(
-            timestamp_as_str.strip()
-        )
-    return timestamp_obj
 
 
 def get_stored_state(
@@ -148,18 +117,18 @@ def generic_web_api_data_etl(
         from_date: datetime = None,
         until_date: datetime = None,
 ):
-    stored_state = get_stored_state(
-        data_config
-    )
-
-    initial_from_date = from_date or stored_state
-    from_date_to_advance = initial_from_date
-    cursor = None
     imported_timestamp = get_current_timestamp_as_string(
         ModuleConstant.DATA_IMPORT_TIMESTAMP_FORMAT
     )
+
+    stored_state = get_stored_state(
+        data_config
+    )
+    initial_from_date = from_date or stored_state
+    from_date_to_advance = initial_from_date
+    cursor = None
     latest_record_timestamp = None
-    offset = 0
+    offset = 0 if data_config.url_manager.offset_param else None
     page_number = 1 if data_config.url_manager.page_number_param else None
     with TemporaryDirectory() as tmp_dir:
         full_temp_file_location = str(
@@ -205,24 +174,31 @@ def generic_web_api_data_etl(
             ):
                 break
 
-        if os.path.getsize(full_temp_file_location) > 0:
-            create_or_extend_table_schema(
-                data_config.gcp_project,
-                data_config.dataset_name,
-                data_config.table_name,
-                full_temp_file_location,
-                quoted_values_are_strings=False
-            )
-
-            load_file_into_bq(
-                filename=full_temp_file_location,
-                table_name=data_config.table_name,
-                auto_detect_schema=False,
-                dataset_name=data_config.dataset_name,
-                project_name=data_config.gcp_project,
-            )
+        load_written_data_to_bq(data_config, full_temp_file_location)
         upload_latest_timestamp_as_pipeline_state(
             data_config, latest_record_timestamp
+        )
+
+
+def load_written_data_to_bq(
+        data_config: WebApiConfig,
+        full_temp_file_location: str
+):
+    if os.path.getsize(full_temp_file_location) > 0:
+        create_or_extend_table_schema(
+            data_config.gcp_project,
+            data_config.dataset_name,
+            data_config.table_name,
+            full_temp_file_location,
+            quoted_values_are_strings=False
+        )
+
+        load_file_into_bq(
+            filename=full_temp_file_location,
+            table_name=data_config.table_name,
+            auto_detect_schema=False,
+            dataset_name=data_config.dataset_name,
+            project_name=data_config.gcp_project,
         )
 
 
@@ -303,7 +279,10 @@ def upload_latest_timestamp_as_pipeline_state(
             data_config.state_file_object_name and
             data_config.state_file_bucket_name
     ):
-        latest_record_date = get_timestamp_as_string(latest_record_timestamp)
+        latest_record_date = datetime_to_string(
+            latest_record_timestamp,
+            ModuleConstant.DEFAULT_TIMESTAMP_FORMAT
+        )
         state_file_name_key = data_config.state_file_object_name
         state_file_bucket = data_config.state_file_bucket_name
         upload_s3_object(
@@ -311,176 +290,3 @@ def upload_latest_timestamp_as_pipeline_state(
             object_key=state_file_name_key,
             data_object=latest_record_date,
         )
-
-
-def get_bq_schema(data_config: WebApiConfig,):
-    if (
-            data_config.schema_file_object_name and
-            data_config.schema_file_s3_bucket
-    ):
-        bq_schema = download_s3_json_object(
-            data_config.schema_file_s3_bucket,
-            data_config.schema_file_object_name
-        )
-    else:
-        bq_schema = None
-    return bq_schema
-
-
-def process_downloaded_data(
-        record_list: list,
-        data_config: WebApiConfig,
-        data_etl_timestamp,
-        file_location,
-        prev_page_latest_timestamp: datetime = None
-):
-    provenance = {
-        data_config.import_timestamp_field_name:
-            data_etl_timestamp
-    }
-    bq_schema = get_bq_schema(data_config)
-    processed_record_list = process_record_in_list(
-        record_list=record_list, bq_schema=bq_schema,
-        provenance=provenance
-    )
-    processed_record_list = iter_write_jsonl_to_file(
-        processed_record_list, file_location
-    )
-    current_page_latest_timestamp = get_latest_record_list_timestamp(
-        processed_record_list, prev_page_latest_timestamp, data_config
-    )
-    return current_page_latest_timestamp
-
-
-def process_record_in_list(
-        record_list,
-        provenance: dict = None,
-        bq_schema=None
-) -> Iterable:
-    for record in record_list:
-        n_record = standardize_record_keys(record)
-        if bq_schema:
-            n_record = filter_record_by_schema(n_record, bq_schema)
-        if provenance:
-            n_record.update(provenance)
-        yield n_record
-
-
-# pylint: disable=inconsistent-return-statements,broad-except,no-else-return
-def standardize_record_keys(record_object):
-    if isinstance(record_object, dict):
-        new_dict = {}
-        for item_key, item_val in record_object.items():
-            new_key = standardize_field_name(item_key)
-            if isinstance(item_val, (list, dict)):
-                item_val = standardize_record_keys(
-                    item_val,
-                )
-            if item_val:
-                new_dict[new_key] = item_val
-        return new_dict
-    elif isinstance(record_object, list):
-        new_list = list()
-        for elem in record_object:
-            if isinstance(elem, (dict, list)):
-                elem = standardize_record_keys(
-                    elem
-                )
-            if elem is not None:
-                new_list.append(elem)
-        return new_list
-
-
-# pylint: disable=inconsistent-return-statements,broad-except,no-else-return
-def filter_record_by_schema(record_object, record_object_schema):
-    if isinstance(record_object, dict):
-        list_as_p_dict = convert_bq_schema_field_list_to_dict(
-            record_object_schema
-        )
-        key_list = set(list_as_p_dict.keys())
-        new_dict = {}
-        for item_key, item_val in record_object.items():
-            if item_key in key_list:
-                if isinstance(item_val, (list, dict)):
-                    item_val = filter_record_by_schema(
-                        item_val,
-                        list_as_p_dict.get(item_key).get(
-                            ModuleConstant.BQ_SCHEMA_SUBFIELD_KEY
-                        ),
-                    )
-                if (
-                        list_as_p_dict.get(item_key)
-                        .get(ModuleConstant.BQ_SCHEMA_FIELD_TYPE_KEY)
-                        .lower() == "timestamp"
-                ):
-                    try:
-                        dateparser.parse(
-                            item_val
-                        )
-                    except BaseException:
-                        item_val = None
-                new_dict[item_key] = item_val
-        return new_dict
-    elif isinstance(record_object, list):
-        new_list = list()
-        for elem in record_object:
-            if isinstance(elem, (dict, list)):
-                elem = filter_record_by_schema(
-                    elem, record_object_schema
-                )
-            if elem is not None:
-                new_list.append(elem)
-        return new_list
-
-
-def get_latest_record_list_timestamp(
-        record_list, previous_latest_timestamp, data_config: WebApiConfig
-):
-    latest_collected_record_timestamp_list = [previous_latest_timestamp]
-
-    for record in record_list:
-
-        if data_config.item_timestamp_key_hierarchy_from_item_root:
-            record_timestamp = parse_timestamp_from_str(
-                get_dict_values_from_hierarchy_as_list(
-                    record,
-                    data_config.item_timestamp_key_hierarchy_from_item_root
-                ),
-                data_config.item_timestamp_format,
-            )
-            latest_collected_record_timestamp_list.append(record_timestamp)
-    latest_collected_record_timestamp_list = [
-        timestamp for
-        timestamp in latest_collected_record_timestamp_list
-        if timestamp
-    ]
-    latest_timestamp = max(
-        latest_collected_record_timestamp_list
-    ) if latest_collected_record_timestamp_list else None
-    return latest_timestamp
-
-
-def get_dict_values_from_hierarchy_as_list(
-        page_data, hierarchy: List[ResponseHierarchyKey]
-):
-    data_value = page_data
-
-    for list_element in hierarchy:
-        data_value = extract_content_from_response(data_value, list_element)
-
-        if not data_value:
-            break
-    return data_value
-
-
-def extract_content_from_response(
-        data_in_response, resp_hierarchy: ResponseHierarchyKey
-):
-    if isinstance(data_in_response, dict) and resp_hierarchy.key:
-        return data_in_response.get(resp_hierarchy.key)
-    elif isinstance(data_in_response, list) and resp_hierarchy.key:
-        return [
-            elem.get(resp_hierarchy.key) for elem in data_in_response
-        ]
-    elif isinstance(data_in_response, dict) and resp_hierarchy.is_variable:
-        return list(data_in_response.items())
