+++ data_pipeline/utils/web_api.py
@@ -1,0 +1,24 @@
+import requests
+from requests.adapters import HTTPAdapter
+# pylint: disable=import-error
+from requests.packages.urllib3.util.retry import Retry
+
+
+def requests_retry_session(
+        retries=10,
+        backoff_factor=0.3,
+        status_forcelist=(500, 502, 504),
+        session=None,
+):
+    session = session or requests.Session()
+    retry = Retry(
+        total=retries,
+        read=retries,
+        connect=retries,
+        backoff_factor=backoff_factor,
+        status_forcelist=status_forcelist,
+    )
+    adapter = HTTPAdapter(max_retries=retry)
+    session.mount('http://', adapter)
+    session.mount('https://', adapter)
+    return session
